server/requirements.txt
asgiref==3.11.1
dj-database-url==3.1.2
Django==6.0.4
django-cors-headers==4.9.0
djangorestframework==3.17.1
djangorestframework-simplejwt==5.3.1
dotenv==0.9.9
gunicorn==25.3.0
packaging==26.2
psycopg2-binary==2.9.12
PyJWT==2.12.1
python-dotenv==1.2.2
sqlparse==0.5.5
whitenoise==6.12.0
